Moved ADC code to correct files
